Extract processing part of FileListener (formerly JobFileProcessor) into JobManager
