+++ lib/job/manager.js
@@ -5,7 +5,7 @@ var mkdirp = require('mkdirp');
 var rimraf = Promise.promisify(require('rimraf'));
 var fs = Promise.promisifyAll(require('fs'));
 
-var JobFileProcessor = require('./file-processor');
+var FileListener = require('./file-listener');
 var JobProcessor = require('./processor');
 
 var serviceLocator = require('../service-locator');
@@ -28,21 +28,64 @@ function JobManager(jobDirectory, tempFilesDir, handlers) {
       this.handlerRegistry.register(handler);
     }.bind(this));
   }
-  this._processor  = new JobProcessor(tempFilesDir);
-  this._fileProcessor = new JobFileProcessor(this._processor, jobDirectory);
+  this._processor = new JobProcessor(tempFilesDir);
+
+  this._fileListener = new FileListener(jobDirectory);
+  this._fileListener.on('file', function(file) {
+    this._processJobFile(file);
+  }.bind(this));
 }
 
 JobManager.prototype.start = function() {
   this._processor.start();
-  this._fileProcessor.processOutstanding();
-  this._fileProcessor.listen();
+  this._fileListener.start();
 };
 
 /**
  * @return {Promise}
  */
 JobManager.prototype.stop = function() {
-  return Promise.all(this._fileProcessor.stop(), this._processor.stop());
+  return Promise.all(this._fileListener.stop(), this._processor.stop());
+};
+
+
+/**
+ * @param {String} filePath
+ * @returns {Promise}
+ */
+JobManager.prototype._processJobFile = function(filePath) {
+  var self = this;
+  return self._readJobFile(filePath)
+    .then(function(jobDescription) {
+      var jobId = path.basename(filePath, '.json');
+      var handler = self.handlerRegistry.get(jobDescription['plugin'], jobDescription['event']);
+      var job = handler.instantiateJob(jobId, jobDescription['data']);
+      return self._processor.processUntilSuccessful(job);
+    })
+    .then(function() {
+      serviceLocator.get('logger').info('Removing ' + filePath);
+      return fs.unlink(filePath);
+    })
+    .catch(function(error) {
+      var logMessage = error.stack || error.message || error;
+      serviceLocator.get('logger').error(logMessage);
+    });
+};
+
+/**
+ * @param {String} filePath
+ * @returns {Promise}
+ */
+JobManager.prototype._readJobFile = function(filePath) {
+  return fs.readFile(filePath)
+    .then(function(content) {
+      return Promise.try(function() {
+        return JSON.parse(content);
+      });
+    })
+    .catch(function(error) {
+      throw new Error('Invalid job file: ' + error.message)
+    });
 };
 
 module.exports = JobManager;
